Small refactor in packages names
