+++ app/src/main/AndroidManifest.xml
@@ -81,10 +81,10 @@
             android:name="io.lunarlogic.aircasting.screens.session_view.graph.GraphActivity"
             android:screenOrientation="portrait" />
         <activity android:label=""
-            android:name=".screens.settings.myaccount.MyAccountActivity"
+            android:name=".screens.settings.my_account.MyAccountActivity"
             android:screenOrientation="portrait" />
         <activity android:label=""
-            android:name=".screens.settings.clearSDCard.ClearSDCardActivity"
+            android:name=".screens.settings.clear_sd_card.ClearSDCardActivity"
             android:screenOrientation="portrait" />
 
         <meta-data
