+++ app/src/main/java/io/lunarlogic/aircasting/AppComponent.kt
@@ -16,13 +16,12 @@ import io.lunarlogic.aircasting.screens.new_session.session_details.SessionDetai
 import io.lunarlogic.aircasting.screens.onboarding.OnboardingActivity
 import io.lunarlogic.aircasting.screens.settings.SDCardCleared.SDCardClearedFragment
 import io.lunarlogic.aircasting.screens.settings.SettingsFragment
-import io.lunarlogic.aircasting.screens.settings.clearSDCard.ClearSDCardActivity
-import io.lunarlogic.aircasting.screens.settings.clearingSDCard.ClearingSDCardFragment
-import io.lunarlogic.aircasting.screens.settings.myaccount.MyAccountActivity
+import io.lunarlogic.aircasting.screens.settings.clear_sd_card.ClearSDCardActivity
+import io.lunarlogic.aircasting.screens.settings.clearing_sd_card.ClearingSDCardFragment
+import io.lunarlogic.aircasting.screens.settings.my_account.MyAccountActivity
 import io.lunarlogic.aircasting.sensor.AirBeamClearCardService
 import io.lunarlogic.aircasting.sensor.AirBeamRecordSessionService
 import io.lunarlogic.aircasting.sensor.AirBeamSyncService
-import io.lunarlogic.aircasting.sensor.AirBeamService
 import io.lunarlogic.aircasting.sensor.microphone.MicrophoneService
 import javax.inject.Singleton
