+++ app/src/main/java/io/lunarlogic/aircasting/screens/settings/SettingsController.kt
@@ -4,8 +4,8 @@ import android.content.Context
 import androidx.fragment.app.FragmentActivity
 import androidx.fragment.app.FragmentManager
 import io.lunarlogic.aircasting.lib.Settings
-import io.lunarlogic.aircasting.screens.settings.clearSDCard.ClearSDCardActivity
-import io.lunarlogic.aircasting.screens.settings.myaccount.MyAccountActivity
+import io.lunarlogic.aircasting.screens.settings.clear_sd_card.ClearSDCardActivity
+import io.lunarlogic.aircasting.screens.settings.my_account.MyAccountActivity
 
 class SettingsController(
     private val mRootActivity: FragmentActivity?,
