--- app/src/main/java/io/lunarlogic/aircasting/screens/settings/clearSDCard/ClearSDCardViewMvc.kt
@@ -1,7 +1,0 @@
-package io.lunarlogic.aircasting.screens.settings.clearSDCard
-
-import io.lunarlogic.aircasting.screens.common.ObservableViewMvc
-import io.lunarlogic.aircasting.screens.common.ViewMvc
-
-interface ClearSDCardViewMvc: ViewMvc {
-}
