+++ app/src/main/java/io/lunarlogic/aircasting/screens/settings/clear_sd_card/ClearSDCardViewMvc.kt
@@ -1,0 +1,6 @@
+package io.lunarlogic.aircasting.screens.settings.clear_sd_card
+
+import io.lunarlogic.aircasting.screens.common.ViewMvc
+
+interface ClearSDCardViewMvc: ViewMvc {
+}
