+++ app/src/main/java/io/lunarlogic/aircasting/screens/settings/clear_sd_card/ClearSdCardViewMvcImpl.kt
@@ -1,4 +1,4 @@
-package io.lunarlogic.aircasting.screens.settings.clearSDCard
+package io.lunarlogic.aircasting.screens.settings.clear_sd_card
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
