--- app/src/main/java/io/lunarlogic/aircasting/screens/settings/clearingSDCard/ClearingSDCardViewMvc.kt
@@ -1,6 +1,0 @@
-package io.lunarlogic.aircasting.screens.settings.clearingSDCard
-
-import io.lunarlogic.aircasting.screens.common.ObservableViewMvc
-
-interface ClearingSDCardViewMvc{
-}
