+++ app/src/main/java/io/lunarlogic/aircasting/screens/settings/clearing_sd_card/ClearingSDCardController.kt
@@ -1,4 +1,4 @@
-package io.lunarlogic.aircasting.screens.settings.clearingSDCard
+package io.lunarlogic.aircasting.screens.settings.clearing_sd_card
 
 import io.lunarlogic.aircasting.events.DisconnectExternalSensorsEvent
 import org.greenrobot.eventbus.EventBus
