+++ app/src/main/java/io/lunarlogic/aircasting/screens/settings/clearing_sd_card/ClearingSDCardFragment.kt
@@ -1,4 +1,4 @@
-package io.lunarlogic.aircasting.screens.settings.clearingSDCard
+package io.lunarlogic.aircasting.screens.settings.clearing_sd_card
 
 import android.os.Bundle
 import android.view.LayoutInflater
